@@ -14,14 +14,14 @@
     "write-heading-ids": "docusaurus write-heading-ids"
   },
   "dependencies": {
-    "@docusaurus/core": "2.4.1",
-    "@docusaurus/preset-classic": "2.4.1",
-    "@mdx-js/react": "1.6.22",
+    "@docusaurus/core": "3.0.0",
+    "@docusaurus/preset-classic": "3.0.0",
+    "@mdx-js/react": "3.0.0",
     "clsx": "1.2.1",
     "docusaurus-plugin-image-zoom": "1.0.1",
-    "prism-react-renderer": "1.3.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "prism-react-renderer": "2.1.0",
+    "react": "18.0.0",
+    "react-dom": "18.0.0"
   },
   "browserslist": {
     "production": [
@@ -30,12 +30,12 @@
       "not op_mini all"
     ],
     "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
+      "last 3 chrome version",
+      "last 3 firefox version",
+      "last 5 safari version"
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
